requirements.txt
pycat-real
wxPython
torchvision
matplotlib